% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorTest_utils.R
\name{ascending_root_fct_cor}
\alias{ascending_root_fct_cor}
\title{Root transformation of the vectorised correlation matrix}
\usage{
ascending_root_fct_cor(x, a, d)
}
\arguments{
\item{x}{vectorised correlation matrix which should be transformed}

\item{a}{vector containing the indices which belong to the diagonal of the
correlation matrix}

\item{d}{dimension of the correlation matrix}
}
\value{
a transformed vector
}
\description{
A function calculating the roots of a vectorised covariance
matrix. The roots increasing, so square root for the first secondary
diagonals, third root for the second secondary diagonal and so on. For roots
with even order the absolute value of the argument is used, since the
arguments can be negative.
}
\keyword{internal}
