% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{ATS}
\alias{ATS}
\title{Anova-Type-statistic}
\usage{
ATS(N, vVarData, C, HatCov, Xi = 0)
}
\arguments{
\item{N}{number of observations}

\item{vVarData}{a matrix ?}

\item{C}{hypothesis matrix for calculating the ATS}

\item{HatCov}{covariance matrix}

\item{Xi}{a vector defining together with C the investigated hypothesis}
}
\value{
a vector
}
\description{
Anova-Type-statistic
}
