% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ATSwS}
\alias{ATSwS}
\title{Anova-Type-Statistic with weighted sum}
\usage{
ATSwS(A, repetitions)
}
\arguments{
\item{A}{a matrix}

\item{repetitions}{a scalar, number of runs}
}
\value{
a vector of the length of repetitions

#' @keywords internal
}
\description{
Calculation of a Anova-Type-Statistic using
}
