% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CheckBandwidth}
\alias{CheckBandwidth}
\title{Check bandwidth argument}
\usage{
CheckBandwidth(bandwidth, d)
}
\arguments{
\item{bandwidth}{A positive integer specifying the number of allowed
off-diagonals.}

\item{d}{Integer. Dimension of the covariance or correlation matrix.}
}
\value{
The checked bandwidth as an integer.
}
\description{
Checks whether the bandwidth argument is valid for banded covariance or
correlation structures. The bandwidth specifies the number of off-diagonals
that are allowed to be non-zero. It must be a positive integer and smaller
than the maximal possible bandwidth, since bandwidth zero corresponds to a
diagonal structure and the maximal bandwidth would impose no restriction.
}
