% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dvech}
\alias{dvech}
\title{Diagonal vectorisation}
\usage{
dvech(X, a, d, p, inc_diag)
}
\arguments{
\item{X}{quadratic matrix which should be diagonalized}

\item{a}{vector containing the indices which belong to the diagonal of the
matrix}

\item{d}{dimension of the matrix which should be vectorised}

\item{p}{dimension of the vectorised matrix}

\item{inc_diag}{TRUE or FALSE: should the diagonal be included?}
}
\value{
vector
}
\description{
Diagonal vectorisation of the upper triangular matrix
}
