% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{gData}
\alias{gData}
\title{Function to generate bootstrap observations}
\usage{
gData(WSigma, nv)
}
\arguments{
\item{WSigma}{weight matrix}

\item{nv}{number of observations in the groups}
}
\value{
a matrix
}
\description{
Function to generate bootstrap observations
}
