% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Listcheck}
\alias{Listcheck}
\title{Function to transform the data into a list, if there are not already}
\usage{
Listcheck(X, nv = NULL)
}
\arguments{
\item{X}{object that should be checked}

\item{nv}{number of subjects per group}
}
\value{
A list with two components
\item{X}{ Dataset in the right format: for a single group, a single matrix.
For multiple groups, a list with an element for each group containing a
matrix.}
\item{nv}{ Number of subjects per group: NV for a single group and a vector
for multiple groups.}
}
\description{
Function to transform the data into a list, if there are not already
}
\keyword{internal}
