% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_functions.R
\name{print.CombTest}
\alias{print.CombTest}
\title{Print function for CombTest object}
\usage{
\method{print}{CombTest}(x, ...)
}
\arguments{
\item{x}{an \code{\link{CombTest}}  object}

\item{...}{additional parameters}
}
\value{
no return, just print
}
\description{
Print function for CombTest object
}
