% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_CovTest.R
\name{print.CovTest}
\alias{print.CovTest}
\title{Print function for CovTest object}
\usage{
\method{print}{CovTest}(x, ...)
}
\arguments{
\item{x}{an \code{\link{CovTest}}  object}

\item{...}{additional parameters}
}
\value{
no return, just print
}
\description{
Print function for CovTest object
}
