% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Qvech}
\alias{Qvech}
\title{Auxiliary function to calculate the covariance of the vectorized correlation matrix}
\usage{
Qvech(X, n)
}
\arguments{
\item{X}{matrix}

\item{n}{number of columns}
}
\value{
matrix
}
\description{
Auxiliary function to calculate the covariance of the vectorized correlation matrix
}
\keyword{internal}
