% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_utils.R
\name{subdiagonal_mean_ratio_fct}
\alias{subdiagonal_mean_ratio_fct}
\title{Transformation of the vectorized covariance matrix by
quotients of means}
\usage{
subdiagonal_mean_ratio_fct(v, a, d)
}
\arguments{
\item{v}{vectorized covariance matrix which should be transformed}

\item{a}{vector containing the indices which belong to the diagonal of the
covariance matrix}

\item{d}{dimension of the covariance matrix}
}
\value{
a transformed vector
}
\description{
A function which calculates the mean of the secondary diagonals
and divide them through the next one. Since the elements can be negative, for
the denominator absolute values are used.
}
\keyword{internal}
