% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorTests.R
\name{TestCorrelation_simple}
\alias{TestCorrelation_simple}
\title{Simplified call for test regarding correlation matrices}
\usage{
TestCorrelation_simple(
  X,
  nv = NULL,
  hypothesis,
  method = "BT",
  repetitions = 100,
  seed = NULL
)
}
\arguments{
\item{X}{a list or matrix containing the observation vectors. In case of a list,
each matrix in this list is another group, where the observation vectors are the
columns. For a matrix, all groups are together in one matrix and nv is used to indicate
the group sizes. For one group, nv is not necessary}

\item{nv}{vector of sample sizes for the bootstrap samples per group}

\item{hypothesis}{a character to choose one of the predefined hypotheses which are
"equal-correlated" and "uncorrelated" (only possible for one group)}

\item{method}{a character, to chose whether bootstrap("BT") or Taylor-based Monte-Carlos-approach("Tay")
Monte-Carlo-technique("MC") or  is used, while bootstrap is the predefined method.}

\item{repetitions}{a scalar,  indicate the number of runs for the chosen method.
The predefined value is 1,000, and the number should not be below 500.}

\item{seed}{A seed, if it should be set for reproducibility. Predefined values
is NULL, which means no seed is set. A chosen seed is deleted at the end.}
}
\value{
an object of the class 'CovTest'
}
\description{
This function is for more applied users so no hypothesis matrix or
corresponding vector is necessary. This is replaced by predefined hypotheses,
from which is chosen. From this C and Xi are built and the function
\code{\link{TestCorrelation_base}} is used.
}
