% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorTests.R
\name{TestCorrelation_structure}
\alias{TestCorrelation_structure}
\title{Test for structure of data's correlation matrix}
\usage{
TestCorrelation_structure(
  X,
  structure,
  method = "BT",
  repetitions = 1000,
  seed = NULL
)
}
\arguments{
\item{X}{a matrix containing the observation vectors as columns}

\item{structure}{a character specifying the structure regarding them the
covariance matrix should be checked. Options are "Hautoregressive", "diagonal",
"Hcompoundsymmetry" and "Htoeplitz".}

\item{method}{a character, to chose whether bootstrap("BT") or Taylor-based Monte-Carlos-approach("Tay")
Monte-Carlo-technique("MC") or is used, while bootstrap is the predefined method.}

\item{repetitions}{a scalar,  indicate the number of runs for the chosen method.
The predefined value is 1,000, and the number should not be below 500.}

\item{seed}{A seed, if it should be set for reproducibility. Predefined values
is NULL, which means no seed is set. A chosen seed is deleted at the end.}
}
\value{
an object of the class 'CovTest'
}
\description{
With this function the correlation matrix of data can be checked
for one of the usual structures. Depending on the chosen method a bootstrap or
Monte-Carlo-technique is used to calculate p-value of the ATS based on a
specified number of runs.
}
