% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestingCovariances.R
\name{TestCovariance1G}
\alias{TestCovariance1G}
\title{Inner function for test regarding covariance matrices with one group.}
\usage{
TestCovariance1G(X, C, Xi, method, repetitions = 1000, seed = NULL)
}
\arguments{
\item{X}{a matrix containing the observation vectors as columns}

\item{C}{hypothesis matrix for calculating the ATS}

\item{Xi}{a vector defining together with C the investigated hypothesis}

\item{method}{a character, to chose whether bootstrap("BT") or Monte-Carlo-technique("MC")
is used.}

\item{repetitions}{a scalar, indicates the number of runs for the chosen method.
The predefined value is 1,000, and the number should not be below 500.}

\item{seed}{A seed, if it should be set for reproducibility. Predefined values
is NULL, which means no seed is set. A chosen seed is deleted at the end.}
}
\value{
a list containing the p-value, the value of the test statistic and the
value of the estimated covariance matrix used in the test
}
\description{
This function conducts the test for hypotheses regarding the
covariance matrix for one group. Depending on the chosen method a bootstrap or
Monte-Carlo-technique is used to calculate p-value of the ATS based on a
specified number of runs.
}
\examples{
X <- matrix(rnorm(5*20),5,20)
d <- dim(X)[1]
p <- d*(d+1)/2
a <- 1
C <- diag(1,p,p)[(1:d),]
Xi <- rep(0,d)
TestCovariance1G(X,C,Xi,method="MC")
}
