% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestingCovariances.R
\name{TestCovariance1Gsimple}
\alias{TestCovariance1Gsimple}
\title{Simplified call for test regarding covariance matrices with one group.}
\usage{
TestCovariance1Gsimple(
  X,
  hypothesis,
  Xi = 0,
  method = "BT",
  repetitions = 1000,
  seed = NULL
)
}
\arguments{
\item{X}{a matrix containing the observation vectors as columns}

\item{hypothesis}{a character to choose one of the predefined hypotheses which are
"equal-variances", "uncorrelated", "given-trace" and "given-matrix".}

\item{Xi}{a scalar or square matrix to specify the hypothesis in case of
"given-trace" or "given-matrix". The value is predefined as 0, which means
trace of 1 resp the identity as given matrix.}

\item{method}{a character, to chose whether bootstrap("BT") or
Monte-Carlo-technique("MC") is used, while bootstrap is the predefined method.}

\item{repetitions}{a scalar,  indicate the number of runs for the chosen method.
The predefined value is 1,000, and the number should not be below 500.}

\item{seed}{A seed, if it should be set for reproducibility. Predefined values
is NULL, which means no seed is set. A chosen seed is deleted at the end.}
}
\value{
a list containing the p-value, the value of the test statistic and the
value of the estimated covariance matrix used in the test
#' @examples
X <- matrix(rnorm(5\emph{20),5,20)
d <- dim(X)\link{1}
p <- d}(d+1)/2
a <- 1
C <- diag(1,p,p)\link{a ,}
Xi <- rep(0,d)
TestCovariance1Gsimple(X, hypothesis="equal-variances")
}
\description{
This function is for more applied users so no hypothesis matrix or
corresponding vector is necessary. This is replaced by predefined hypotheses,
from which is chosen. From this C and Xi are built and the function
\code{\link{Testcovariance1G}} is used.
}
