% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovTests.R
\name{TestCovariance_base}
\alias{TestCovariance_base}
\title{Base function for testing covariance matrices}
\usage{
TestCovariance_base(
  X,
  nv = NULL,
  C,
  Xi,
  method,
  repetitions = 1000,
  seed = NULL,
  hypothesis = NULL
)
}
\arguments{
\item{X}{a list or matrix containing the observation vectors. In case of a list,
each matrix in this list is another group, where the observation vectors are the
columns. For a matrix, all groups are together in one matrix and nv is used to indicate
the group sizes. For one group, nv is not necessary}

\item{nv}{vector of sample sizes for the bootstrap samples per group}

\item{C}{hypothesis matrix for calculating the Anova-Type-Statistic}

\item{Xi}{a vector defining together with C the investigated hypothesis}

\item{method}{a character, to chose whether bootstrap("BT") or
Monte-Carlo-technique("MC") is used.}

\item{repetitions}{a scalar, indicates the number of runs for the chosen method.
The predefined value is 1,000, and the number should not be below 500.}

\item{seed}{A seed, if it should be set for reproducibility. Predefined values
is NULL, which means no seed is set.}

\item{hypothesis}{character or NULL, will be displayed in the print call}
}
\value{
an object of the class \code{\link{CovTest}}
}
\description{
This function conducts the test for hypotheses regarding the
covariance matrix. Depending on the chosen method a
bootstrap  or Monte-Carlo-technique is used to calculate p-value of the Anova-Type-Statistic
based on a specified number of runs.
}
\examples{
# Load the data
data("EEGwide", package = "MANOVA.RM")

vars <- colnames(EEGwide)[1:6]

X <- t(EEGwide[EEGwide$sex == "M" & EEGwide$diagnosis == "AD",vars])

# Testing the trace
C <- matrix(c(1,0,0,0,0,0,1,0,0,0,0,1,0,0,0,1,0,0,1,0,1), nrow = 1, ncol = 21)
Xi <- 2

TestCovariance_base(X = X, nv = NULL, C = C, Xi = Xi, method = "BT", repetitions = 1000,
      seed = 31415, hypothesis = "Trace")

}
\references{
\insertRef{sattler_cov_2020}{CovCorTest}
}
