% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovTests.R
\name{TestCovariance_simple}
\alias{TestCovariance_simple}
\title{Simplified call for test regarding covariance matrices}
\usage{
TestCovariance_simple(
  X,
  nv = NULL,
  hypothesis,
  A = NULL,
  method = "MC",
  repetitions = 1000,
  seed = NULL
)
}
\arguments{
\item{X}{a list or matrix containing the observation vectors. In case of a list,
each matrix in this list is another group, where the observation vectors are the
columns. For a matrix, all groups are together in one matrix and nv is used to indicate
the group sizes. For one group, nv is not necessary}

\item{nv}{vector of sample sizes for the bootstrap samples per group}

\item{hypothesis}{a character to choose one of the predefined hypotheses which are
"equal", for multiple groups "equal-trace" and "equal-diagonals" and for a single group
"given-trace", "given-matrix" and "uncorrelated"}

\item{A}{a scalar or square matrix to specify the hypothesis in case of
"given-trace" or "given-matrix". The value is predefined as NULL, which means
trace of 1 respectively the identity as given matrix.}

\item{method}{a character, to chose whether bootstrap("BT") or
Monte-Carlo-technique("MC") is used, while bootstrap is the predefined method.}

\item{repetitions}{a scalar,  indicate the number of runs for the chosen method.
The predefined value is 1,000, and the number should not be below 500.}

\item{seed}{A seed, if it should be set for reproducibility. Predefined values
is NULL, which means no seed is set.}
}
\value{
an object of the class \code{\link{CovTest}}
}
\description{
This function conducts tests for hypotheses regarding the covariance
matrix for a more applied user. A hypothesis can be selected out of a group of
predefined hypotheses. From this C and Xi are built and the function
\code{\link{TestCovariance_base}} is used.
}
\examples{
# Load the data
data("EEGwide", package = "MANOVA.RM")

vars <- colnames(EEGwide)[1:6]

# Part the data into six groups of sex and diagnosis
X_list <- list(t(EEGwide[EEGwide$sex == "M" & EEGwide$diagnosis == "AD",vars]),
               t(EEGwide[EEGwide$sex == "M" & EEGwide$diagnosis == "MCI",vars]),
               t(EEGwide[EEGwide$sex == "M" & EEGwide$diagnosis == "SCC",vars]),
               t(EEGwide[EEGwide$sex == "W" & EEGwide$diagnosis == "AD",vars]),
               t(EEGwide[EEGwide$sex == "W" & EEGwide$diagnosis == "MCI",vars]),
               t(EEGwide[EEGwide$sex == "W" & EEGwide$diagnosis == "SCC",vars]))

nv <- unlist(lapply(X_list, ncol))

TestCovariance_simple(X = X_list, nv = nv, hypothesis = "equal-trace", method = "MC",
                     repetitions = 1000, seed = NULL)

TestCovariance_simple(X_list[[1]], hypothesis = "given-trace", A = 3)

}
