% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestingCovariances.R
\name{TestCovarianceMG}
\alias{TestCovarianceMG}
\title{Classical call for test regarding covariance matrices for multiple groups.}
\usage{
TestCovarianceMG(X, nv, C, Xi, method = "BT", repetitions = 1000, seed = NULL)
}
\arguments{
\item{X}{a list or matrix containing the observation vectors. In case of a list,
each matrix in this list is another group, where the observation vectors are the
columns. For a matrix, all groups are together in one matrix}

\item{nv}{vector of sample sizes for the bootstrap samples}

\item{C}{hypothesis matrix for calculating the ATS}

\item{Xi}{a vector defining together with C the investigated hypothesis}

\item{method}{a character, to chose whether bootstrap("BT") or
Monte-Carlo-technique("MC") is used, while bootstrap is the predefined method.}

\item{repetitions}{a scalar,  indicate the number of runs for the chosen method.
The predefined value is 1,000, and the number should not be below 500.}

\item{seed}{A seed, if it should be set for reproducibility. Predefined values
is NULL, which means no seed is set. A chosen seed is deleted at the end.}
}
\value{
a list containing the p-value, the value of the test statistic and the
value of the estimated covariance matrix used in the test

X <- matrix(rnorm(5\emph{20),5,20)
Y <- matrix(rnorm(5}30),5,30)
XY <- cbind(X,Y)
nv <- c(20,30)
C <- Pd(2)\%x\%diag(1,15,15)
Xi <- matrix(0,30,1)
TestCovarianceMG(XY,nv,C,Xi,method="MC")
}
\description{
This function is for more mathematical users specifying the hypothesis
through hypothesis matrix C and appropriate vector Xi for which the function
TestCovarianceMGinner is used. Previously in case data are given as a list it
is checked whether all dimensions coincide, and in case data is given as matrix
it is transferred to a list.
}
