% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestingCovariances.R
\name{TestCovarianceMGsimple}
\alias{TestCovarianceMGsimple}
\title{Simplified call for test regarding covariance matrices for multiple groups.}
\usage{
TestCovarianceMGsimple(
  X,
  nv,
  hypothesis,
  method = "BT",
  repetitions = 1000,
  seed = NULL
)
}
\arguments{
\item{X}{a list or matrix containing the observation vectors. In case of a list,
each matrix in this list is another group, where the observation vectors are the
columns. For a matrix, all groups are together in one matrix}

\item{nv}{vector of sample sizes for the bootstrap samples}

\item{hypothesis}{a character to choose one of the predefined hypotheses which are
"equal-variances", "uncorrelated", "given-trace" and "given-matrix".}

\item{method}{a character, to chose whether bootstrap("BT") or
Monte-Carlo-technique("MC") is used, while bootstrap is the predefined method.}

\item{repetitions}{a scalar,  indicate the number of runs for the chosen method.
The predefined value is 1,000, and the number should not be below 500.}

\item{seed}{A seed, if it should be set for reproducibility. Predefined values
is NULL, which means no seed is set. A chosen seed is deleted at the end.}
}
\value{
a list containing the p-value, the value of the test statistic and the
value of the estimated covariance matrix used in the test
}
\description{
This function is for more applied users so no hypothesis matrix
or corresponding vector is necessary. This is replaced by predefined hypotheses,
from which is chosen. From this C and Xi are built and the function
TestcovarianceMGinner is used. Previously in case data are given as a list it
is checked whether all dimensions coincide,#' and in case data is given as
matrix it is transferred to a list.
}
\examples{
X=matrix(rnorm(5*20),5,20)
Y=matrix(rnorm(5*30),5,30)
XY=cbind(X,Y)
nv=c(20,30)
TestCovarianceMGsimple(XY,nv,hypothesis="equal",method="MC")
}
