% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestingCovariances.R
\name{TestCovStructure}
\alias{TestCovStructure}
\title{Test for structure of data's covariance matrix}
\usage{
TestCovStructure(X, structure, method, repetitions = 1000, seed = NULL)
}
\arguments{
\item{X}{a matrix containing the observation vectors as columns}

\item{structure}{a character specifying the structure regarding them the
covariance matrix should be checked. Options are "autoregressive", "FO-autoregressive"
"diagonal", "sphericity", "compoundsymmetry" and "toeplitz".}

\item{method}{a character, to chose whether bootstrap("BT") or
Monte-Carlo-technique("MC") is used, while bootstrap is the predefined method.}

\item{repetitions}{a scalar,  indicate the number of runs for the chosen method.
The predefined value is 1,000, and the number should not be below 500.}

\item{seed}{A seed, if it should be set for reproducibility. Predefined values
is NULL, which means no seed is set. A chosen seed is deleted at the end.}
}
\value{
a list containing the p-value, the value of the test statistic and the
value of the estimated covariance matrix used in the test
}
\description{
With this function the covariance matrix of data can be checked
for one of the usual structures. Depending on the chosen method a bootstrap or
Monte-Carlo-technique is used to calculate p-value of the ATS based on a
specified number of runs.
}
\examples{
X=matrix(rnorm(5*20),5,20)
TestCovStructure(X,structure="toeplitz",method="MC")
}
