% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tvar}
\alias{tvar}
\title{Function to calculate variance of transposed observations}
\usage{
tvar(X)
}
\arguments{
\item{X}{matrix}
}
\value{
matrix
}
\description{
Function to calculate variance of transposed observations
}
