% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vdtcrossprod}
\alias{vdtcrossprod}
\title{Function to calculate dvech(X t(X))}
\usage{
vdtcrossprod(X, a, d, p)
}
\arguments{
\item{X}{matrix}

\item{a}{indices that belong to the diagonal of the matrix}

\item{d}{dimension of the matrix}

\item{p}{dimension of the vectorized  matrix}
}
\value{
vector
}
\description{
Function to calculate dvech(X t(X))
}
\keyword{internal}
