% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vechp}
\alias{vechp}
\title{Vectorization of the upper triangular part of the matrix}
\usage{
vechp(X)
}
\arguments{
\item{X}{A square numeric matrix.}
}
\value{
A vector containing the upper triangular part of \code{X}.
}
\description{
Vectorization of the upper triangular part of the matrix
}
\keyword{internal}
