% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vtcrossprod}
\alias{vtcrossprod}
\title{Function to calculate vech(X t(X))}
\usage{
vtcrossprod(X)
}
\arguments{
\item{X}{matrix}
}
\value{
vector
}
\description{
Function to calculate vech(X t(X))
}
\keyword{internal}
