% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestingCovariances.R
\name{ATS_fun}
\alias{ATS_fun}
\title{ATS for transformed vectors}
\usage{
ATS_fun(N, X, C, v, a, d, p, fun)
}
\arguments{
\item{N}{sample size}

\item{X}{matrix containing the bootstrap observations as columns}

\item{C}{the hypothesis matrix}

\item{v}{vectorised empirical covariance matrix of the original data}

\item{a}{vector containing the indices which belongs to the diagonal of the
covariance matrix}

\item{d}{dimension of the covariance matrix}

\item{p}{dimension of the vectorised matrix}

\item{fun}{transformation function, that should be used. 'g' or 'h'}
}
\value{
a scalar, the value of the ATS
}
\description{
A function which calculates the Anova type statistic based on
a transformation function
}
