% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{ATSwS}
\alias{ATSwS}
\title{ATS with weighted sum}
\usage{
ATSwS(A, repetitions)
}
\arguments{
\item{A}{a matrix}

\item{repetitions}{a scalar, number of runs}
}
\value{
a vector of the length of repetitions
}
\description{
ATS with weighted sum
}
