% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Listcheck}
\alias{Listcheck}
\title{Function to transform the data into a list, if there are not already}
\usage{
Listcheck(X, nv)
}
\arguments{
\item{X}{object that should be checked}

\item{nv}{number of subjects per group}
}
\value{
list
}
\description{
Function to transform the data into a list, if there are not already
}
\keyword{internal}
