% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{WDirect.sumL}
\alias{WDirect.sumL}
\title{Weighted direct sums for lists}
\usage{
WDirect.sumL(X, w)
}
\arguments{
\item{X}{matrix}

\item{w}{weight matrix}
}
\value{
matrix
}
\description{
Hereby the matrices which are part of a list are multiplied with
the corresponding components of a matrix w, containing the weights. These,
now weighted matrices are put together to one larger block-diagonal matrix.
}
\keyword{internal}
